The mean value of the Euclidean Nearest-Neighbor Distance of patches in each class(vector data)

This function allows you to calculate the mean value of Euclidean Nearest-Neighbor Distance among patches of one class in a categorical landscape in vector data format Euclidean Nearest-Neighbor Distance means the distance from a patch edge to the nearest neighbouring patch belonging to the same class.

Usage

vm_c_enn_mn(landscape, class_col)

Arguments

landscape

the input landscape image,

class_col

the name of the class column of the input landscape

Value

the returned calculated mean value of each class is in column "value", and this function returns also some important information such as level, class number and metric name. Moreover, the "id" column, although it is just NA here at class level. we need it because the output struture of metrics at class level should correspond to patch level one by one, and then it is more convinient to combine metric values at different levels and compare them.

Examples

vm_c_enn_mn(vector_landscape, "class")
#> # A tibble: 3 × 5
#>   level class id    metric value
#>   <chr> <chr> <chr> <chr>  <dbl>
#> 1 class 1     NA    enn_mn    NA
#> 2 class 2     NA    enn_mn    NA
#> 3 class 3     NA    enn_mn    NA